Add initial support for running in a container.
This commit is focused on building and running the
container in a local docker environment. Parameters
were added to the go program to allow us to adjust
the number of registration retries and the time between
each retry, looking forward to the need to wait for a
delay in our container's external network connectivity.

An example env file, example-container.env, has been provided
to show how to pass parameters to our container. Commentary in
example-container.env describes how to build a local Docker
container and then run it using the env file.

Documentation for each important environment variable can be found
in the container's startup script, container-startup.sh.