We can now handle multiple rooms registered by us.
If we start/stop/start and occasionally use different
room names, then we will accumulate multiple registered
rooms. We now query for the list of rooms after we register
and remember the correlation between room id and name so
that we can reference the correct room names in our banter
back to the players.

As part of this work several unused structs and/or struct fields
were removed.

Also, checkpoint() was changed to only add a newline after printing.
Previously it also added a newline at the start of printing.